feat: export register and command
