+++ src/watcher.ts
@@ -2,7 +2,7 @@ import path from 'path';
 import chokidar from 'chokidar';
 import assert from 'assert';
 import { EventEmitter } from 'events';
-import { TsGenerator, TsGenConfig, TsHelperConfig, default as TsHelper } from './';
+import { TsGenerator, TsGenConfig, TsHelperConfig, default as TsHelper } from './core';
 import * as utils from './utils';
 import { loadGenerator } from './generator';
 import d from 'debug';
